requirements.txt
selenium>=4.0.0
webdriver-manager>=3.8.0
google-generativeai>=0.3.0